***BROKEN*** Don't use CP to configure VM pagetable bases
*** THIS IS A QUICK HACK ***
It works most of the time but it does not synchronize properly and may
cause random GPU pagefaults.

On PS4 the CP ring can't write to the pagetable registers (there seems
to be a hardcoded access control in SRBM which rejects writes to that
register range coming from the GRMB block, where CP issues its writes
from).

This patch just writes them from the CPU, but does not do the necessary
synchronization so it is wrong.

We may be able to write to the registers from the SDMA ring. If so,
maybe a better approach would be to have a barrier to synchronize the
rings, then have SDMA do it instead, and have CP wait on that.